LayoutTests/css3/blending/isolation-isolate-simple-expected.txt

PASS 0 is 0
PASS 0 is 0
PASS successfullyParsed is true

TEST COMPLETE
